+++ app/(auth)/login/page.tsx
@@ -1,4 +1,3 @@
-import type { Metadata } from "next";
 import LoginPage from "./LoginPage";
 
 export default function Login() {
