+++ components/ui/submit-button.tsx
@@ -8,18 +8,20 @@ interface SubmitButtonProps {
   label: string;
   pendingLabel: string;
   className?: string;
+  disabled?: boolean;
 }
 
 export function SubmitButton({
   label,
   pendingLabel,
   className,
+  disabled,
   ...props
 }: SubmitButtonProps) {
   const { pending } = useFormStatus();
 
   return (
-    <Button type="submit" disabled={pending} className={className} {...props}>
+    <Button type="submit" disabled={pending || disabled} className={className} {...props}>
       {pending ? pendingLabel : label}
     </Button>
   );
